refactor: Update import statements in tabular.py and test_datasets_schemas.py
